pyproject.toml
[project]
name = "auementations"
version = "0.1.0"
description = "A composable audio augmentation library with Hydra configuration support"
readme = "README.md"
authors = [
    { name = "Christopher Jacoby", email = "christopher@antarestech.com" }
]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "hydra-core>=1.3.0",
    "hydra-zen>=0.13.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
torch = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "torch-audiomentations>=0.11.0",
]
pedalboard = [
    "pedalboard>=0.9.0",
]
all = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "torch-audiomentations>=0.11.0",
    "pedalboard>=0.9.0",
]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
]

[tool.hatch.build.targets.wheel]
packages = ["auementations"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
